// assets/Levels/types.ts
import { DIRECTION_ENUM, ENTITY_TYPE_ENUM, ENTITY_STATE_ENUM, TILE_TYPE_ENUM } from '../Enums'

export interface ITile {
  src: number | null
  type: TILE_TYPE_ENUM | null
}

export interface ILevel {
  mapInfo: Array<Array<ITile>>
  player: IEntity
  enemies: Array<IEntity>
  spikes: Array<ISpikes>
  bursts: Array<IEntity>
  door: IEntity
}

export interface IEntity {
  x: number
  y: number
  direction: DIRECTION_ENUM
  state: ENTITY_STATE_ENUM
  type: ENTITY_TYPE_ENUM
}

export interface ISpikes {
  x: number
  y: number
  type: ENTITY_TYPE_ENUM
  count: number
}